+++ migrations/02_drop_col_tid.sql
@@ -1,0 +1,2 @@
+ALTER TABLE positions
+DROP COLUMN tid;